Add initial code for turning VMX on
Right now, we have just added a single intrinsic, which uses the VMXON
instruction to put the active CPU into VMX host mode. In the long term,
this should probably not be its own intrinsic, but rather something
that's automatically done at boot-up during SVA initialization.

(On a non-SVA system, the hypervisor would use VMXON/VMXOFF when
starting/stopping the hypervisor application; since only one hypervisor
can have control of the CPU at a time, this allows other hypervisors to
be used non-concurrently on the same system. But under SVA, this will be
unnecessary, since SVA will have sole responsibility for managing the
processor's VMX features, and hypervisor software will go through it.
There will never be a need for SVA to release control of the VMX
features to any other software.)

To test this code, add some simple kernel code (e.g. a system call or
character device) that calls the new intrinsic. (That is not included in
this respository because it's in the separately-cloned FreeBSD90 repo.
The system call is trivial, so I have not committed it to FreeBSD90.)