Check CR0 and CR4 settings for permissibility in VMX operation
If these are set wrong, the VMXON instruction will fail. Intel provides
four MSRs that specify which bits must be 0 and 1 in CR0 and CR4,
respectively, for VMXON to proceed.

This is basically a sanity check: if the bits are set wrong, it's
because the OS is doing something broadly incompatible with VMX (e.g.,
not using paging). We do not expect this to be the case in normal
operation. However, it's important to check this for the sake of
forward-compatibility, in case future processors' implementations of VMX
have stricter requirements.