Verify that kernel isn't lying to us when it gives us SVA frames
When we ask the kernel to give SVA a frame of memory for its own use, we
are expecting it to have no outstanding mappings by which it can access
it, including removing it from its own direct map.

However, since this is the responsibility of kernel code, we can't trust
it to do what it's supposed to. (E.g. a rootkit could replace the
provideSVAMemory() function with one that keeps frames in the direct
map, allowing it to read ghost memory.) Thus, whenever we call
provideSVAMemory(), we need to confirm that these conditions are in fact
true.

Now, all SVA memory is vetted before it's placed into SVA's frame cache.
This covers ghost memory as well as non-ghost frames used internally by
SVA and taken from the frame cache (e.g., various VMX structures).