Eliminate race condition in code protecting SVA/ghost memory
Since we are vetting frames for safety (i.e. ensuring there are no rogue
mappings to them) when we first obtain them from the OS, they need to
remain protected the whole time they are sitting in SVA's frame cache.

We ensure this by setting the frame type to PG_SVA when it is first
obtained from the OS. However, the ghost memory code also needed to be
updated to take this into account: when de-allocating ghost frames, it
should set the type back to PG_SVA, not PG_UNUSED (which does not
protect against rogue mappings).

Pre/postconditions have been added to the alloc/free_frame() functions
to make these "rules" clear for any future use of SVA's secure memory
system.