Start writing VM entry/exit context switch code
This is turning out to be a fair amount of work, so I'm going to commit
what I've done so far even though it's not done yet.

Although the new intrinsic (sva_launchvm()) isn't complete yet, it
compiles fine, and isn't called anywhere else (so it shouldn't affect
existing functionality).

Segmentation might be mostly gone in 64-bit, but there's still enough
of it left that I had to spend the better part of today reading all
about it to figure out how to save/restore that pesky task register and
the descriptor table registers... :-|

Pro(??) tip: today I discovered that the AMD manual is written very much
like a textbook, making it a great resource alongside the Intel manual
when you really want to figure out how to *use* a processor feature,
instead of just an (incredibly detailed) summation of what each of its
functions do...