Add sva_resumevm() intrinsic
This is the last of the intrinsics to be implemented.

The next step is to test the functionality of the launch/resumevm
intrinsics by actually creating a "hello world" VM. That will take place
outside of this repository, since it is FreeBSD kernel code. (I'll be
creating a private fork of John's FreeBSD90 repository for this. We
probably will not want to upstream my changes to his primary FreeBSD90
repository, because this is just a simple "smoke test" for the VMX
intrinsics. What we really want to do in the big picture is to port
the BHyVe hypervisor to SVA; unfortunately, that doesn't exist until
FreeBSD 10.)