Set up EPT hierarchy for "hello world" VM
Adds a (temporary) intrinsic that sets up a simple EPT hierarchy mapping
a single frame into the guest-physical address space.

We will write some test code to the beginning of this frame and set the
guest's RIP to start there on VM launch.

For the "hello world" VM, we can get away without having to set up a
full guest-side paging hierarchy by running the guest in unpaged
(64-bit) mode. Thus we only need to worry about the
host-physical-to-guest-physical mapping (and not a
guest-physical-to-guest-virtual mapping in addition to that). This is
possible thanks to the "unrestricted guest" feature available in newer
Intel processors (including our Broadwell development hardware); older
processors only supported running guests in paged mode.