Move debug intrinsic to debug.c, split exported interface out of vmx.h
An intrinsic for printing MSRs has been moved to debug.c where it joins
many of similar ilk.

This necessitated moving a bunch of VMX-related constants/declarations
out of vmx.c, which was a good idea anyway. However, for the sake of
clean software engineering, these should not be exported as part of the
SVA-VMX interface; they are only needed by SVA to support its access to
the raw hardware VMX interface. Thus, I followed the example of
mmu.c/mmu_intrinsics.c and created separate headers for the
internal/external aspects of the VMX code.