Change guest-mode test program to read from stack vs. registers
This is needed since we don't yet have an interface to allow the caller
of sva_launch/resumevm() (i.e. the hypervisor) to
initialize/save/restore the GPRs across VM entry/exit. Right now, we
just save/restore the host registers and leave the registers
"uninitialized" (actually filled with host values!) on VM entry.

This will clearly need to be fixed to support more complex VMs (and
shouldn't be too hard to do), but it's not on the critical path to
getting a minimal "hello world" VM up and running.

(Note: this matches commit 93d8e2d in the ethanjohnson89/FreeBSD90
repository.)