Fix reading and parsing of TR descriptor
I had reversed the order of the fields written out by the "sgdt" and
"sidt" instructions; this was causing me to read a garbage value for the
GDT base.

Additionally, I was doing pointer arithmetic wrong when calculating the
address of the TR descriptor within the GDT. (I'd neglected to consider
that I was working with a uint16_t pointer and thus all math is in units
of 2 bytes.)

Now, we successfully make it through to the VMLAUNCH instruction.
However, the VM launch is failing because (the error code says) I've set
one of the myriad control fields wrong. I'm debugging that now...