Use paging in the "hello world" guest
My brilliant plan to make setting up the guest's memory environment
simpler by running with paging disabled was foiled by AMD's design
choice to not support running without paging in 64-bit mode.

I believe this to be (at least one of) the reasons why I was getting the
"guest state invalid" error on VM launch. (There may be other reasons,
but we have to fix this first to find out.)

Alternatively, we could have kept the guest unpaged and switched it to
32-bit mode (which is supported), but that would've been at least as
much work, since running in 64-bit mode allows us to not have to worry
about (relatively) as many details in setting up the segmentation
hardware consistently.

To-do in the FreeBSD90 repository:
- Set guest CR3 to point to the guest-side paging hierarchy we're now
  creating.

- Adjust guest CR0/CR4 as appropriate to reflect that we're using
  paging.