Improve VMX debug printouts
We're still getting the "guest state invalid" error, even after
switching to using paging in the guest (which should take care of one of
those invalid pieces of guest state, namely, that CR0 was configured
incorrectly for 64-bit mode).

We now print human-readable names for VMCS fields when reading/writing
them. This makes it easier to stare at the configuration over and over
again to try to find which checks are (still!!!) failing.