Saved host segment selectors must have RPL=0, TI=0
This is an error that Bochs caught. It's a real error according to the
Intel manual (which Bochs follows to the letter) but, oddly enough, the
real hardware didn't get stuck on this.

However, it did draw my attention to processor behavior I
was previously unaware of: the DPL will be hard-set to 0 in the
actual in-processor segment registers on VM exit, regardless of what
they were on VM entry. (There's no way to override this.)

This is probably not actually a problem, but I'd like to be sure of
that. I spent some time thinking about this and recorded my (very
rough) thoughts in the comments committed here. I'm probably
overthinking this and will realize why when I get the chance to read the
access-control sections of the Paging chapter of the manual a bit more
carefully.