Don't use guest-physical addresses over 46 bits
Now that we've fixed the errors that Bochs was catching that weren't
showing up on the (less picky) real hardware, we're actually seeing the
real cause(s) of the "invalid guest state" error.

The first error was "Invalid Guest CR3". It turns out the root causes is
that I was exceeding the max-physical-address limit in my guest-physical
EPT address mapping.

According to CPUID leaf 0x80000008, our development hardware is limited
to 46-bit physical addresses. A closer look at the Intel manual
confirmed that this limitation applies to *guest*-physical addresses as
well as host-physical.

We were attempting to map pages into guest-physical addresses starting
at 0xf00dbeef0000, which has bits set up through #47.

Bochs's simulated machine has an even tighter maximum physical address
of 40 bits - but its CPUID instruction lies about this and reports 46.
Go figure. (Examining the Bochs source confirms that it's actually
limited to 40 bits, and that's what its VMX checks enforce.)

This commit also fixes a small (but significant) typo in the code that
was incorrectly using host-physical instead of guest-physical addresses
in the guest's page tables. (Not to be confused with the EPT page
tables. That was where I messed up...)

Corresponds to commit 21fbf9d in the FreeBSD90 repo:
https://github.com/ethanjohnson89/FreeBSD90/commit/21fbf9d12a57876e6d61e7b93810331f9f1a1b76