Make guest-virtual addresses canonical (and now it works! almost)
Since our mapped guest-virtual addresses have bit 47 set to 1, bits
63-48 must also be 1. These addresses are in the "upper half" of the
64-bit virtual address space as laid out by the x86-64 architecture
(i.e., 46 bits' worth is defined at each of the top and bottom of the
space, with everything in-between undefined until future processors
expand the linear-address size, allowing kernels to use the standard
trick of "supervisor owns top half" in the meantime).

I feel silly now for not catching this when I first set up the address
space...I "knew" this. :-)

We don't actually need to make any changes to the guest page tables,
since the higher bits aren't actually used for indexing in 4-level
paging. (If they were, we wouldn't have to set them this way...)

Meanwhile, GOOD NEWS: with this fix, we now have a successful VM launch!
But we're not out of the woods yet: the VM is page-faulting on its very
first instruction fetch. Working on that next...