Fix CPUID assembly mistake (and also print linear-address width)
I was reading the wrong register as output from CPUID, so the result was
bogus.

My previous assertion that Bochs was "lying" in reporting its
physical-address width was based on this bogus printout and therefore
wrong. Bochs is actually reporting exactly what it should. Both it, and
the physical hardware we are developing on, have a physical-address
limit of 40 bits.