Fix bit math fail -> VM launch success! :-D
I goofed on the bit mask to align the guest-virtual address of the
mapped region to a 1 GB boundary. That left some reserved bits in the
page table entry set to 1, which was triggering a page fault.

With this fixed, we now can successfully launch our simple "hello world"
VM. :-D