Fixed a severe bug in the SFI pass.
The extra AND instruction in SFI-arith was discussed before but not
recognized as a bug. However, it actually *is* a bug. Consider an
address within kernel map that starts with 0xffff ff80. The AND
instruction with mask 0xffff fd00 would clear its 41st (0 indexed) bit
since 0xd = 1101b. The pointer would then be seen as referring to SVA
DMAP. Luckily, the OR instruction does not change the pointer value and
that is why the kernel has not broken yet.