Extend simple test VM to test VMRESUME
After the first VM exit, we adjust the guest's RIP to skip over the HLT
instruction that triggered the exit, and resume. It performs another
simple computation, pushes the result to the stack, and triggers a VM
exit using the CPUID instruction.

Also see companion commit recently pushed to the kernel repo:
https://github.com/ethanjohnson89/FreeBSD90/commit/82553ea5cc2a9d1973ebbf4b1e46a35bfbcb7507