Save guest GPRs on VM exit so we can print them
This will give us *much* better visibility into what's going on in VM
exit situations.

Currently this is being done in a kind of ugly way (saving the guest
GPRs into the host state structure because it's the only structure we
have a pointer to after VM exit). This is not sufficient for supporting
actual save/restore of guest registers across VM entry/exit, something
that we will need to support in the very near future. I need to do a
little more thinking about how to do that - there's nothing conceptually
difficult about it, but the interactions between the VM entry/exit
inline assembly and host data structures are a bit "interesting".