Save/restore host RFLAGS so interrupts are re-enabled on VM exit
It turns out that RFLAGS is cleared unconditionally on VM exit, which
means that interrupts are blocked until we re-enable them.

(Aside:)
I hadn't noticed this before because the SYSRET back to userspace when
our syscall finishes restores the user-space value of RFLAGS, which (of
course) has IF = 1, i.e. interrupts get re-enabled at that time. Since
our code was returning to user-space within seconds of the last VM exit,
interrupts never remained blocked long enough to cause major problems.
However, it is possible to "notice" this if you press Scroll Lock fast
enough while the syscall is still executing - the screen won't actually
freeze until the return to user space. (I'm guessing this is due to the
keyboard interrupt being held up, though it's possible that's normal
behavior of Scroll Lock during a syscall. I haven't been able to test if
the behavior's changed after the fix since I'm not on the physical
console today and my remote-access setup has just enough latency to keep
me from pressing Scroll Lock in the short window of time.)