Fix very silly mistake in copying code from host to guest
In order to give the guest a program to run, we need to copy some code
into the guest's address space before the first VMLAUNCH. For our simple
test VM, this is a statically defined (fancy way of saying "hardcoded")
program.

In my profound wisdom, I had deemed the strcpy() function to be an
appropriate means of installing the guest code. I profoundly regretted
this when I tried to use some "move immediate" instructions to load
small constants in the guest program. The zero bytes in the immediate
values caused strcpy() to stop copying, truncating the guest code.

This caused all manner of hilariously bizarre side effects, most notably
the inevitable page fault that occurs when one tries to execute the
opcode 0x0000 (which is what was left in place of the rest of the
truncated guest code), a.k.a. "addb %al, (%rax)", with a small integer
in RAX. I spent more time than I care to admit debugging this, as the
root cause was very non-obvious.

We now install the guest's program in a much more sane way. We include
the guest program as an inline assembly block within SVA's code, and
have SVA memcpy() it into the guest's address space.

(Hat-tip to Isaac Richter for suggesting this idea.)