Fix minor inline assembly usage mistake
I read up a bit more on inline assembly syntax and realized I was doing
something slightly wrong in the block starting at line 1974. It was
behaving correctly, but technically the compiler could've made unsafe
assumptions about the code. Specifically, I was writing to memory
through pointers without declaring memory as clobbered. Adding the
clobber declaration would've been one way to fix it, but I ended up
going with a cleaner way (which gives the compiler the choice of whether
to use a register or memory operand - though it chooses memory here
anyway).

The other inline assembly block changed in this commit wasn't wrong, but
I realized I'd overconstrained the operands, limiting the compiler to
using memory stores instead of register-to-register moves if
appropriate. (It actually ends up compiling it the same way, but that
could change if the surrounding code, compiler version, or optimization
level changes.)

Lastly, I removed the #if 0'd blocks that I added in the last commit - I
don't actually want to keep them around, I just wanted them preserved
for posterity in Git in case I want to refer later to how I did things
in them. (In particular, the code to pretty-print an entire page's
contents is handy utility code to have "on a shelf".)