Fix AT&T vs. Intel assembly syntax fail
I've been reading too much Intel manual lately, so I got the operand
order mixed up on this instruction. This was causing the code to
erroneously save the guest registers into the host_state structure
(instead of the VM descriptor).

Hiliariously (but dangerously), the code almost "just worked", for all
the wrong reasons, which is why I missed it when I tested the code
before the last commit. Because the reversed instruction overwrote the
active_vm pointer with a pointer to the host_state structure - and the
printf() code that displays the guest's registers after every VM exit
uses that pointer to access them - the printouts were reading from the
same (wrong) place that the registers were being written to, by
pseudo-coincidence. So at a quick glance, it looked like everything was
working correctly; but a closer look revealed that some of the host's
registers were getting corrupted due to their structure being
overwritten.

Isn't C fun? :-)