Add intrinsic support for specifying initial guest state
The client (kernel/hypervisor) can now specify the initial GPR values
and RIP/RSP settings when calling sva_allocvm().

This commit also restructures some code and data structures to support
upcoming intrinsics that will allow the hypervisor to get/set the guest
state (e.g. to handle a VM exit that requires emulation of behavior that
a guest can't do on the real hardware).

Also see this commit in the FreeBSD90 repo which updates our toy
hypervisor to use the new interface:
https://github.com/ethanjohnson89/FreeBSD90/commit/0150070990e03c7db7110f115e76c413205fc120