Set unchanging VMCS fields in SVA instead of hypervisor
There are a few VMCS fields that always need to be set a certain way,
either for security reasons or because they relate to features that we
don't support yet in SVA (or don't ever intend to support because
they're legacy features). We should therefore set them ourselves instead
of relying on the hypervisor to set them correctly.