Improve formatting/commenting in sva_remove_page()
No functional changes to code. (I renamed a variable to make it more
self-explanatory, but it's the same to the computer.)

These changes came from my reading of the code to understand it before
expanding it to support EPT for VMX. I'm committing them separately so
that those subsequent functional changes can be more clearly seen in the
commit history.

I don't really understand what the #ifdef SVA_ASID_PG section does (it
is sadly lacking in comments) but I improved its formatting so I could
actually read it.