Fix inconsistent indentation in sva_create_dmap()
No code changes. Done as a byproduct of reading the code to try to
figure out what it does.

(This code originally made inconsistent use of soft and hard tabs which
made it very difficult to read in my editor. It now follows the
two-space soft-tab convention used predominantly throughout the rest of
SVA.)