Implement MMU checks for EPT mapping updates
The new checks are largely integrated into the existing non-EPT code
since most of the logic is shared.

As a side effect of this, I significantly refactored the internal helper
function pt_update_is_valid(), which is where most of the MMU checks
happen. The logic is now much clearer, which made it easier to handle
EPT special cases without making the code convoluted.

The refactoring doesn't seem to have broken anything, although I haven't
tested it with ghost memory (e.g. attempting to have the OS map a ghost
frame or ghost PTP).