Fix bugs in EPT MMU checks
Two minor bugs appeared in testing with the toy hypervisor:

1. I'd oversimplified the bit logic in the is-present check for EPT
mappings. Among other bad effects, such as hiding bug #2 below, this
specifically manifested itself in refcounts not being incremented when
EPT mappings are created.

2. I neglected to include the EPT case in pt_update_is_valid()'s
determination of whether it is looking at a last-level or higher-level
mapping. This manifested itself as a false positive on one of the
security checks when attempting to create an EPT mapping, because it was
impossible to create the last-level mapping to the actual physical
memory (SVA thought it was non-last-level and expected it to point to a
declared PTP).