Fix off-by-two error in page refcounts + PTP undeclare consistency issue
Page refcounts are *supposed* to include the references in both direct
maps (the kernel's and SVA's). (I confirmed this with Xiaowan, and this
expectation is also reflected by various SVA checks which look for
things like "count <= 2".)

However, for whatever reason, SVA was never actually incrementing the
refcounts to reflect the existence of the DMAPs. This created a security
loophole wherein a lot of checks would be skipped because SVA didn't
realize the OS still held mappings to a frame. (e.g. you could get away
with undeclaring a PTP while it's still referenced by an upstream PTP)

Once this was fixed, it uncovered another bug: when a PTP is undeclared
that still contains valid mappings, we need to decrement the refcounts
for those mappings to reflect the fact that they are thus implicitly
removed. Failure to do this would lead to refcounts remaining
artificially inflated, causing false positives on some security checks.
(e.g. undeclaring an L2 PTP pointing to an L1, then undeclaring the L1,
would lead to SVA panicking because the L1 appeared to still be mapped
somewhere)

This bug was less likely (but still possible) to be triggered before the
off-by-two error was fixed, because this situation had to happen twice
before it'd fail a check.