Include top-level extended page table page references in refcounts
We now include the reference in a VM's EPTP (equivalent of CR3) in the
refcount for the top-level page-table page it points to. This closes a
loophole wherein the system software could undeclare a level-4 PTP while
it's still actively in use by a VM.

A similar loophole needs to be closed for CR3 and regular paging (still
to-do).