Remove extraneous TLB flush in sva_mm_load_pgtable()
We were saving/reloading CR3 to clear non-global TLB entries after
ensuring that the secmem PML4 entry was in place following a page table
switch. However, this was already happening when we loaded CR3 to
activate the new page tables, so by slightly rearranging the code (to
make sure the secmem entry is in place *before* loading the new page
table), we only need to do it once.

This has the potential to improve performance since TLB flushes are an
expensive operation. (Plus it makes the code simpler/clearer.)