Remove my_getVirtual() hack
I wrote this out of an abundance of caution to make sure my VMX code
would work correctly before I actually knew what was going on with the
two DMAPs. Now that I know how they work it's clear I just want to call
getVirtualSVADMAP() directly. (The VMX code in general doesn't attempt
to behave correctly in the #ifndef SVA_DMAP case, which seems to be
vestigial since we've had the SVA DMAP for a while now and I don't see
any reason why we'd want to go back.)