Fix subtle but scary typo in ASID switch function
Today I learned that clang doesn't throw an error or even warn you when
you specify a completely nonexistent variable name as an input/output to
an inline assembly block. I'm not sure how it thought it could compile
that but it must have been wrong. (Nothing operational was actually
broken, but that was totally by accident because this code was *beyond*
undefined... :-|)

It turns out the inline assembly isn't even needed here because we have
helper functions for that.