Set MSR load/store VMCS fields unconditionally in lieu of VMCS check
There is only one safe value, known at compile time, that SVA can permit
these fields to be set to - so, rather than having a check that forces
the system software to do this "boilerplate" step correctly, SVA can
just do it itself.

In general, we try not to expose features unsupported by SVA to the
system software at all (if that's feasible, as it is here).