Fix security hole in TLB invalidation
This code is responsible for ensuring that the OS cannot gain
"back-door" write access to a page that has been appropriated for use as
a PTP or ghost memory by exploiting stale TLB entries that it neglected
to flush.

It was correct at the time it was written, but had not been updated to
reflect the addition of SVA's DMAP (which changed the baseline for
refcounts). That re-opened the security hole it was supposed to prevent.