Flush EPT TLB entries when doing global TLB flushes
When a page is appropriated by SVA for secure use as a PTP or ghost
memory, we need to do a global TLB flush to ensure that the OS has no
stale mappings that can still access it.

We now also take into account the possibility of stale *EPT* mappings
that might point to the page. (EPT mappings are not flushed by the
standard methods of flushing a TLB; the INVEPT instruction must be used
instead.)

Incidental to this, we now only do a global TLB flush when SVA is about
to actually take control of a page. Previously, we did it whenever the
last non-DMAP reference to a page was removed. That strategy was safe
(because removing all non-DMAP references is an enforced prerequisite to
handing over a page to SVA for secure use), but confusing and suboptimal
for performance. We are now only doing an expensive global TLB flush
when it's actually needed for security.

This should have a positive performance impact on SVA, potentially a
significant one (though I don't know how much it'll come out to in the
real world).