Include config.h before vmx_intrinsics.h to avoid compile error
For some reason, the compiler wasn't recognizing #ifdef MPX to be in
effect in vmx_intrinsics.h, even though it *was* recognizing it in
vmx.c, which led to a compile error because the MPX-related code in
vmx.c was trying to use identifiers which were supposed to be defined by
corresponding code in vmx_intrinsics.h.

The problem went away when I made sure we were #including <sva/config.h>
before any other SVA headers. I don't know why this makes a difference,
since config.h doesn't change the MPX define in any way (it only uses it
with and #ifdef)...but it's working now, so I'm not complaining.