Remove more console spam so we can test on real hardware
Apparently, even having just *one* thing printed to the screen when
handling an external interrupt is enough to delay things long enough to
get another external interrupt...and so on and so on, preventing us from
ever getting anything executed in the guest.

With all printouts removed on the external-interrupt-exit-handling path,
we can handle the (in my observations) merely *one* interrupt that
typically arrives during our short test quickly enough to have plenty of
time to finish the test.