Revert commit 3ee79d3 ("Added dummy FPU code to populate the FPU")
This was (I think) causing kernel panics for me when I tried to run the
test suite on Marquez in multi-user mode.

(I'm not 100% sure that this code was the cause of the kernel panics,
but they went away when I removed it. As @kdharsee previously indicated
in a comment on commit 3ee79d3, this wasn't meant to be committed
anyway.)