Add padding to guest state structure for side-channel protection
This guest state structure is the only thing in SVA protected memory
that we touch in the small window of time between switching to the OS
cache partition and VM entry (and vice versa on VM exit). We don't care
if the guest state structure itself is subject to side-channel attacks
because it's already under control of the system software, but we need
to make sure there's nothing sensitive adjacent to it in the cache line
that'll also get pulled into the OS partition.