Clarify/clean up cache partition switching code
We are using an undocumented preliminary implementation of the Intel CAT
feature on our development hardware (marquez.cs.rochester.edu with an
i7-6700) that differs slightly from the Intel manual's description of
how to use the MSR that sets the class of service. To avoid confusion
going forward, we now have a comment in icat.h explaining this.

I also added some comments to the partition-switching assembly in the
trap/syscall handlers and cleaned up some confusing assembly logic that
I think was slightly buggy in one place (it was spilling registers to
the application's GS segment but then doing a SWAPGS to the kernel's GS
before restoring them; if I'm interpreting the code correctly, the
SWAPGS should happen after restoring the spilled registers).