Disable FPU printouts on VM entry/exit (needed on real hardware)
When we're testing on real hardware (not inside a VM), we can't have
anything printed to screen on the critical path for handling VM exits
due to external interrupts. The printout itself takes long enough for
another external interrupt to come in, getting us stuck in an cycle of
handling external interrupts indefinitely without being able to make any
progress in the guest.

See also commit d1eb01f where we previously addressed this before we
added the FPU code.