Implement VMCS read checks
Similar to sva_writevmcs(), sva_readvmcs() now uses a whitelist of
known-safe VMCS fields that we can allow the system software to read
without risk of leaking sensitive information. An attempt to read a
control not on the list will be rejected.

In the future, as we add controls to the whitelist to support more VMX
features, we may want to be more selective about some fields, e.g.
sanitize them to remove sensitive information in particular subfields
rather than reject the read outright. For now, though, the checks are
pretty simple because our toy hypervisor is only reading fields that (as
far as I can tell) are indeed unconditionally safe.