Bug fix: load initial VMCS values in sva_loadvm() instead of run_vm()
This fixes an issue where sva_readvmcs() would return uninitialized
garbage values, and sva_writevmcs() would write values that wouldn't
take effect, if executed in the time window between creating a VM with
sva_allocvm() and running it for the first time with sva_launchvm().

Basically, we were loading the initial values into the VMCS the first
time the VM was run, when we really should've been doing it the first
time it was loaded, because that's the earliest opportunity for us to do
so (and, thus, the earliest opportunity for the system software to use
sva_read/writevmcs()).