Handle ghost PTPs separately when removing outstanding mappings
Since commit 58da429, SVA has (correctly) done two things it didn't
previously do:
1. Include the two direct maps in physical frame refcounts
2. Properly decrement refcounts for "stranded" valid mappings within a
   PTP that is undeclared

However, these fixes didn't handle ghost memory correctly, for two
reasons:
1. Ghost PTPs use their own refcount system separate from regular
   mappings created by the OS using the MMU intrinsics.
2. Most of the ghost-memory-related code (ghost alloc/free functions,
   ghost fault, etc.) was written for a baseline refcount of 0, not 2,
   which (when it was written) correctly reflected the fact that we were
   neglecting to count the DMAPs in refcounts.

This commit fixes both of these issues, thereby resolving a crash that
manifested itself when the system would re-use a ghost PTP that was used
by a previous process that left a mapping to that PTP "stranded" when it
undeclared all its PTPs on program exit. When that occurred, we would
incorrectly remove the ghost PTP mapping using the method we would use
for regular mappings, thus incorrectly decrementing its refcount in the
physical frame metadata. When the second process to use the PTP did
this, the frame count would get low enough to trigger a sanity-check
assert and cause a panic.