Revert bypass of secmem frame cache for unchecked benchmarks
It turns out this actually made things slower, because in the particular
benchmark for which we made this change (VM create/destroy), we are
staying within the buffer afforded by the frame cache. It was actually
15% slower to go directly to the OS every time we wanted a frame. This
seems to be because SVA's frame cache queue is a lighter-weight
mechanism than FreeBSD's pmap allocator (which is unsurprising,
considering the SVA frame cache implementation is dead simple).

I've determined that there are no secmem-related security checks on the
critical path taken by the create/destroy-VM benchmark, so it's OK for
us to allocate VMCS frames in SVA VM memory in the baseline, just like
we normally would. There's nothing about the critical-path code this
takes that's any different from what a reasonable non-SVA-based
hypervisor might do.