Add intrinsic for debugging GS.BASE issues
We were running into some trouble with Xen clobbering SVA's GS.BASE in
kernel mode, so we added this intrinsic which can be called from various
places in Xen's and SVA's code where we suspect things might have gotten
messed up.

This intrinsic is for debugging purposes and thus is *not* currently
called by any other committed code. (We temporarily added calls to it in
various places in the codebase during the debugging process.)

Although we've fixed the particular bug that inspired its creation, I'm
committing this intrinsic since it may be useful in the future
(particularly as we port Xen's VMX support and get into situations where
Xen might be trying to restore GS.BASE on VM exit).