Change VMID type from size_t to int
Initially making this size_t was a poor design choice on my part. It's
always been int in the design as presented in the paper, and now the
code reflects that.

This needs to be a signed type because we want to be able to return
negative error codes from sva_allocvm().