Add transitional intrinsics to support porting Xen's VMX code
Right now we've handed off responsibility for VMCS allocation and
deallocation to SVA (i.e. Xen is calling sva_allocvm()/sva_freevm()
instead of its own code for allocating and freeing VMCS memory), but
since Xen is not yet using SVA intrinsics through the rest of the VM's
lifecycle, it needs to continue to be able to work with the VMCS
directly. To support that, I've added two intrinsics in the "cheater's
code" section that let Xen convert between SVA VM IDs and VMCS physical
address pointers.

(These intrinsics should actually be harmless from a security
perspective, since once we change our SFI instrumentation to actually do
enforcement, it won't matter if Xen has a pointer to a VMCS - it won't
be able to touch it anyway. But that would also make these intrinsics
kind of useless... :-))

Also disabled some more sva_allocvm() code in the #ifdef XEN case since
Xen is passing null pointers right now for sva_allocvm()'s arguments
specifying the initial state of the VM. This is a temporary hack that
will be removed when more of the VM lifecycle is in SVA's hands.