[typecounts] Create new frame refcount system
The primary difference from the old system is the addition of type
counts: In addition to its ordinary reference count, each frame has a
"type count" which counts the number of uses which force the frame's
current type. In order for a frame to change type, its type count must
be 0. Also, when taking a reference to a frame, you now need to specify
the type as which you want to use the frame. If that type does not match
the frame's current type, an error is raised.