Remove some unused and vestigial code
Of note, one of the things being removed is the automatic (un)setting of
page table pages in the kernel's direct map as read-only: it is now the
kernel's responsibility to do this before declaring them. This not only
simplifies SVA, it also means that we don't have to trust (and therefore
don't have to verify) the kernel's direct map to actually be a direct
map.