Don't use `make`'s implicit rules
The implicit rules (despite what the docs say) put `CFLAGS` before
`CPPFLAGS` on the compiler command line, which means that you cannot
create language-specific overrides for things set in `CPPFLAGS`. This is
very annoying, as `CPPFLAGS` is used for anything which goes through the
C preprocessor (which may include things like assembly and linker
scripts in addition to C and C++ code).