Add an intrinsic to create a new icontext
Previously, the only way to make a new icontext was to clone the current
one (i.e. `fork()`). This new intrinsic allows for a new icontext to be
created with no relation to the current one, and allows its user-space
entry point to be set. This allows for "spawning" new threads without
duplicating the context of the current thread.